server/requirements.txt
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
haversine==2.2.0
python-dateutil==2.8.1
requests==2.24.0
urllib3==1.25.9
gunicorn==20.0.4
whitenoise==5.1.0